requirements.txt
dspy>=2.6.27
faster-whisper>=1.1.1
graphviz>=0.20.3
matplotlib>=3.10.3
mlflow>=3.1.0
networkx>=3.5
openai-whisper @ git+https://github.com/openai/whisper.git
pydub>=0.25.1
pyvis>=0.3.2
streamlit>=1.45.1
torch>=2.7.1
fastapi
uvicorn
pydantic
python-dotenv